try try

+++ scripts/tomcat_stop.sh
@@ -2,4 +2,6 @@
 
 service tomcat7 stop
 
-rm -r /var/lib/tomcat7/webapps/spring-mvc-demo-0.0.1-SNAPSHOT/
+if [ -d /var/lib/tomcat7/webapps/spring-mvc-demo-0.0.1-SNAPSHOT/ ]; then
+  rm -r /var/lib/tomcat7/webapps/spring-mvc-demo-0.0.1-SNAPSHOT/
+fi
